Print elements of a linked lists

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,6 +1,4 @@
 #include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
 #include "lists.h"
 
 /**
@@ -12,7 +10,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	int count;
+	size_t count;
 
 	count = 0;
 	while (h != 0)
